refactor: decompose syncEngine + adapters into focused modules
syncEngine.js (943→~450): extract syncManifest, syncGallery,
syncConflict, syncSerialization into sync/ subdirectory.

gdriveAdapter.js (1117→3 facade): extract gdriveAuth, gdriveFolders,
gdriveFiles into gdrive/ subdirectory.

dropboxAdapter.js (643→~380+auth): extract dropboxAuth into
dropbox/ subdirectory.

All public API preserved via facade re-exports.
