+++ src/core/services/syncEngine.js
@@ -1,10 +1,21 @@
-import { db, getSyncDeletedEntries, clearSyncDeletedEntry } from '@/utils/db.js';
-import { encryptForSync, decryptFromSync, hasSyncKey, getSyncKey } from '@/core/services/crypto/keyManager.js';
-import { isSyncIncludingApiKeys } from '@/core/config/ProviderProfiles.js';
-import { showToast } from '@/core/states/toastState.js';
+import { hasSyncKey } from '@/core/services/crypto/keyManager.js';
+
+import { entryKey, clone, getDeviceId, readLocalManifestV2, writeLocalManifestV2, collectSingletonEntries, buildLocalManifestV2, readCloudManifestV2 } from './sync/syncManifest.js';
+import { dataUrlToBinary, computeImageHash, guessImageExt, computeBinaryHash, applyGalleryEntriesForChar, removeGalleryEntryFromChar } from './sync/syncGallery.js';
+import { needsConflict, getLocalConflictEntity, getConflictName, resolveConflict as _resolveConflict } from './sync/syncConflict.js';
+import { computeSyncHash, encryptEntity, decryptEntity, getLocalCharacter, getLocalCharacterWithImages, getLocalPersona, getLocalChat, deleteCloudFileIfExists, readCloudEntityByEntry, applyCloudEntry } from './sync/syncSerialization.js';
+
+export { ENTITY_TYPES, CLOUD_BASE, cloudPath, getDeviceId };
+export { computeSyncHash } from './sync/syncSerialization.js';
+export { entryKey, MANIFEST_VERSION } from './sync/syncManifest.js';
+export { needsConflict, getLocalConflictEntity, getConflictName } from './sync/syncConflict.js';
+export { getLocalCharacter, getLocalCharacterWithImages, getLocalPersona, getLocalChat } from './sync/syncSerialization.js';
+export { computeImageHash, guessImageExt, dataUrlToBinary, computeBinaryHash } from './sync/syncGallery.js';
+export { buildManifest, readLocalManifestV2, writeLocalManifestV2, buildLocalManifestV2, readCloudManifestV2, collectSingletonEntries } from './sync/syncManifest.js';
+export { applyCloudEntry, readCloudEntityByEntry, encryptEntity, decryptEntity, deleteCloudFileIfExists } from './sync/syncSerialization.js';
+export { applyGalleryEntriesForChar, removeGalleryEntryFromChar } from './sync/syncGallery.js';
 
 const CLOUD_BASE = '/Glaze';
-
 const MAX_SYNC_PAYLOAD_BYTES = 30 * 1024 * 1024;
 
 const ENTITY_TYPES = {
@@ -55,56 +66,12 @@ function galleryCloudPath(charId, imgId, imgExt) {
     return `${CLOUD_BASE}/gallery/${charId}/${imgId}.${imgExt}`;
 }
 
-async function dataUrlToBinary(dataUrl) {
-    const res = await fetch(dataUrl);
-    return await res.arrayBuffer();
-}
-
-async function computeImageHash(dataUrl) {
-    const binary = await dataUrlToBinary(dataUrl);
-    const digest = await crypto.subtle.digest('SHA-256', binary);
-    return Array.from(new Uint8Array(digest)).map(b => b.toString(16).padStart(2, '0')).join('');
-}
-
-function guessImageExt(dataUrl) {
-    const match = dataUrl.match(/^data:image\/([\w+]+)/);
-    if (!match) return 'png';
-    const raw = match[1].toLowerCase();
-    if (raw === 'jpeg') return 'jpg';
-    return raw;
-}
-
-function generateDeviceId() {
-    const stored = localStorage.getItem('gz_sync_device_id');
-    if (stored) return stored;
-    const id = Date.now().toString(36) + Math.random().toString(36).substr(2, 8);
-    localStorage.setItem('gz_sync_device_id', id);
-    return id;
-}
-
-function getDeviceId() {
-    return localStorage.getItem('gz_sync_device_id') || generateDeviceId();
-}
-
-async function encryptEntity(data, key) {
-    if (!key) return data;
-    return encryptForSync(data, key);
-}
-
-async function decryptEntity(encrypted, key) {
-    if (!key) return encrypted;
-    if (!encrypted.iv || !encrypted.data) return encrypted;
-    return decryptFromSync(encrypted, key);
-}
-
-const MANIFEST_VERSION = 2;
-
-function entryKey(type, id) {
-    return `${type}:${id}`;
-}
-
-function clone(obj) {
-    return JSON.parse(JSON.stringify(obj));
+function getBreakdownBucket(type) {
+    if (type === ENTITY_TYPES.CHARACTER) return 'characters';
+    if (type === ENTITY_TYPES.PERSONA) return 'personas';
+    if (type === ENTITY_TYPES.CHAT) return 'chats';
+    if (type === ENTITY_TYPES.GALLERY) return 'gallery';
+    return 'settings';
 }
 
 async function getCloudVerificationCandidate(adapter) {
@@ -113,11 +80,7 @@ async function getCloudVerificationCandidate(adapter) {
         const path = file.path_display || file.path || '';
         return path !== cloudPath(ENTITY_TYPES.MANIFEST);
     });
-
-    if (!candidate) {
-        return null;
-    }
-
+    if (!candidate) return null;
     return adapter.download(candidate.path_display || candidate.path);
 }
 
@@ -133,518 +96,11 @@ export async function cloudHasData(adapter) {
 export async function verifyCloudKey(adapter, key) {
     const candidate = await getCloudVerificationCandidate(adapter);
     if (!candidate) return true;
-
     const parsed = JSON.parse(candidate.data);
     if (!_encryptionEnabled) return true;
     if (!parsed.iv || !parsed.data) return true;
-
     await decryptEntity(parsed, key);
     return true;
-}
-
-export async function buildManifest(lastSync, deviceId) {
-    return {
-        version: MANIFEST_VERSION,
-        deviceId: deviceId || getDeviceId(),
-        lastSync: lastSync || 0,
-        createdAt: Date.now(),
-        entries: {}
-    };
-}
-
-async function computeSyncHash(data) {
-    const normalized = JSON.stringify(data ?? null);
-    const bytes = new TextEncoder().encode(normalized);
-    const digest = await crypto.subtle.digest('SHA-256', bytes);
-    return Array.from(new Uint8Array(digest)).map(b => b.toString(16).padStart(2, '0')).join('');
-}
-
-async function migrateV1ManifestToV2(adapter, v1Manifest) {
-    const cloudFiles = await listAllFiles(adapter);
-    const entries = {};
-
-    for (const file of cloudFiles) {
-        const filePath = file.path_display || file.path || '';
-        const cloudModified = file.serverModified ? new Date(file.serverModified).getTime() : Date.now();
-
-        let type = null;
-        let id = null;
-
-        if (filePath.startsWith(`${CLOUD_BASE}/characters/`)) {
-            type = ENTITY_TYPES.CHARACTER;
-            id = filePath.replace(`${CLOUD_BASE}/characters/`, '').replace(/\.(enc|json)$/, '');
-        } else if (filePath.startsWith(`${CLOUD_BASE}/personas/`)) {
-            type = ENTITY_TYPES.PERSONA;
-            id = filePath.replace(`${CLOUD_BASE}/personas/`, '').replace(/\.(enc|json)$/, '');
-        } else if (filePath.startsWith(`${CLOUD_BASE}/chats/`)) {
-            type = ENTITY_TYPES.CHAT;
-            id = filePath.replace(`${CLOUD_BASE}/chats/`, '').replace(/\.(enc|json)$/, '');
-        } else if (filePath.startsWith(`${CLOUD_BASE}/`)) {
-            const fileName = filePath.replace(`${CLOUD_BASE}/`, '').replace(/\.(enc|json)$/, '');
-            if (fileName === 'lorebooks') { type = ENTITY_TYPES.LOREBOOKS; id = 'lorebooks'; }
-            else if (fileName === 'api_presets') { type = ENTITY_TYPES.API_PRESETS; id = 'api_presets'; }
-            else if (fileName === 'theme_presets') { type = ENTITY_TYPES.THEME_PRESETS; id = 'theme_presets'; }
-            else if (fileName === 'theme_state') { type = ENTITY_TYPES.THEME_STATE; id = 'theme_state'; }
-            else if (fileName === 'local_storage') { type = ENTITY_TYPES.LOCAL_STORAGE; id = 'local_storage'; }
-            else { type = ENTITY_TYPES.LOCAL_STORAGE; id = fileName; }
-        }
-
-        if (!type || !id) continue;
-
-        const entryKeyStr = entryKey(type, id);
-        entries[entryKeyStr] = {
-            type,
-            id,
-            path: filePath,
-            updatedAt: cloudModified,
-            hash: null,
-            deleted: false
-        };
-    }
-
-    const migrated = {
-        version: MANIFEST_VERSION,
-        deviceId: v1Manifest.deviceId || getDeviceId(),
-        lastSync: v1Manifest.lastSync || 0,
-        createdAt: v1Manifest.createdAt || Date.now(),
-        entries
-    };
-
-    await adapter.upload(cloudPath(ENTITY_TYPES.MANIFEST), JSON.stringify(migrated));
-    return migrated;
-}
-
-async function readCloudManifestV2(adapter) {
-    const manifest = await readManifest(adapter);
-    if (!manifest) return null;
-    if (manifest?.version === MANIFEST_VERSION && manifest?.entries) {
-        return manifest;
-    }
-    if (manifest?.version === 1 && !manifest?.entries) {
-        return migrateV1ManifestToV2(adapter, manifest);
-    }
-    return null;
-}
-
-async function readLocalManifestV2() {
-    return (await db.get('gz_sync_manifest_v2')) || null;
-}
-
-async function writeLocalManifestV2(manifest) {
-    await db.set('gz_sync_manifest_v2', manifest);
-}
-
-async function collectSingletonEntries() {
-    const singletons = [
-        { type: ENTITY_TYPES.LOREBOOKS, id: 'lorebooks', data: await db.get('gz_lorebooks') },
-        { type: ENTITY_TYPES.API_PRESETS, id: 'api_presets', data: await db.get('gz_api_connection_presets') },
-        { type: ENTITY_TYPES.THEME_PRESETS, id: 'theme_presets', data: await db.get('gz_theme_presets') },
-        { type: ENTITY_TYPES.THEME_STATE, id: 'theme_state', data: await db.get('gz_theme_active_preset') }
-    ];
-
-    const lsData = {};
-    const lsKeys = [
-        'silly_cradle_presets', 'silly_cradle_current_preset_id', 'gz_preset_connections',
-        'regex_scripts', 'gz_active_persona_id', 'gz_persona_connections',
-        'gz_lang', 'gz_theme', 'gz_chat_padding_lr', 'gz_force_mobile_layout',
-        'gz_battery_saver_ui',
-        // Legacy LLM runtime
-        'api-endpoint', 'api-max-tokens', 'api-context',
-        'gz_api_provider', 'gz_api_endpoint_normalized',
-        'gz_api_temp', 'gz_api_topp', 'gz_api_stream',
-        'gz_api_auto_hide_images', 'gz_api_auto_hide_images_n',
-        'gz_api_request_reasoning', 'gz_api_reasoning_start', 'gz_api_reasoning_end', 'gz_api_reasoning_effort',
-        'gz_api_omit_temperature', 'gz_api_omit_top_p', 'gz_api_omit_reasoning', 'gz_api_omit_reasoning_effort',
-        'gz_api_connect_timeout', 'gz_api_stream_timeout',
-        // Embeddings (non-sensitive)
-        'gz_embedding_use_same', 'gz_embedding_target', 'gz_embedding_scan_depth',
-        'gz_embedding_threshold', 'gz_embedding_top_k', 'gz_embedding_max_chunk_tokens',
-        'gz_embedding_enabled',
-        // Image Gen (non-sensitive)
-        'gz_imggen_enabled', 'gz_imggen_api_type', 'gz_imggen_endpoint', 'gz_imggen_model',
-        'gz_imggen_size', 'gz_imggen_quality', 'gz_imggen_aspect_ratio', 'gz_imggen_image_size',
-        'gz_imggen_naistera_model', 'gz_imggen_naistera_aspect_ratio',
-        'gz_imggen_naistera_send_char_avatar', 'gz_imggen_naistera_send_user_avatar',
-        'gz_imggen_routmy_model', 'gz_imggen_routmy_aspect_ratio', 'gz_imggen_routmy_image_size',
-        'gz_imggen_routmy_quality', 'gz_imggen_routmy_send_char_avatar', 'gz_imggen_routmy_send_user_avatar',
-        'gz_imggen_image_context_enabled', 'gz_imggen_image_context_count',
-        'gz_imggen_additional_refs',
-        // Provider profiles (metadata only; the JSON with keys is conditional below)
-        'gz_active_llm_profile_id', 'gz_service_profile_map', 'gz_provider_profiles_migrated',
-        'gz_sync_include_api_keys'
-    ];
-    const includeKeys = isSyncIncludingApiKeys();
-    if (includeKeys) {
-        lsKeys.push('api-key', 'api-model', 'gz_provider_profiles', 'gz_imggen_api_key', 'gz_imggen_naistera_api_key', 'gz_imggen_routmy_api_key');
-    }
-    for (const k of lsKeys) {
-        const v = localStorage.getItem(k);
-        if (v !== null) lsData[k] = v;
-    }
-    singletons.push({ type: ENTITY_TYPES.LOCAL_STORAGE, id: 'local_storage', data: Object.keys(lsData).length > 0 ? lsData : null });
-    return singletons;
-}
-
-async function buildLocalManifestV2() {
-    const manifest = await buildManifest(Date.now(), getDeviceId());
-    const previousManifest = await readLocalManifestV2();
-    const previousEntries = previousManifest?.entries || {};
-    manifest.entries = {};
-
-    const characters = await db.getAll('characters');
-    for (const char of characters) {
-        if (!char?.id) continue;
-        const { images, ...charNoImages } = char;
-        const hash = await computeSyncHash(charNoImages);
-        const previousEntry = previousEntries[entryKey(ENTITY_TYPES.CHARACTER, char.id)];
-        manifest.entries[entryKey(ENTITY_TYPES.CHARACTER, char.id)] = {
-            type: ENTITY_TYPES.CHARACTER,
-            id: char.id,
-            path: cloudPath(ENTITY_TYPES.CHARACTER, char.id),
-            updatedAt: char.updatedAt || previousEntry?.updatedAt || Date.now(),
-            hash,
-            deleted: false
-        };
-
-        if (Array.isArray(images)) {
-            for (const img of images) {
-                if (!img?.id || !img?.src) continue;
-                const imgExt = guessImageExt(img.src);
-                const imgHash = await computeImageHash(img.src);
-                const galleryKey = entryKey(ENTITY_TYPES.GALLERY, `${char.id}:${img.id}`);
-                const prevImgEntry = previousEntries[galleryKey];
-                manifest.entries[galleryKey] = {
-                    type: ENTITY_TYPES.GALLERY,
-                    id: `${char.id}:${img.id}`,
-                    charId: char.id,
-                    imgId: img.id,
-                    path: galleryCloudPath(char.id, img.id, imgExt),
-                    ext: imgExt,
-                    updatedAt: prevImgEntry?.hash === imgHash && !prevImgEntry?.deleted
-                        ? prevImgEntry.updatedAt
-                        : Date.now(),
-                    hash: imgHash,
-                    deleted: false
-                };
-            }
-        }
-    }
-
-    const personas = await db.getAll('personas');
-    for (const persona of personas) {
-        if (!persona?.id) continue;
-        const hash = await computeSyncHash(persona);
-        const previousEntry = previousEntries[entryKey(ENTITY_TYPES.PERSONA, persona.id)];
-        manifest.entries[entryKey(ENTITY_TYPES.PERSONA, persona.id)] = {
-            type: ENTITY_TYPES.PERSONA,
-            id: persona.id,
-            path: cloudPath(ENTITY_TYPES.PERSONA, persona.id),
-            updatedAt: persona.updatedAt || previousEntry?.updatedAt || Date.now(),
-            hash,
-            deleted: false
-        };
-    }
-
-    const chats = await db.getChats();
-    for (const [charId, chatData] of Object.entries(chats || {})) {
-        const hash = await computeSyncHash(chatData);
-        const previousEntry = previousEntries[entryKey(ENTITY_TYPES.CHAT, charId)];
-        manifest.entries[entryKey(ENTITY_TYPES.CHAT, charId)] = {
-            type: ENTITY_TYPES.CHAT,
-            id: charId,
-            path: cloudPath(ENTITY_TYPES.CHAT, charId),
-            updatedAt: chatData?.updatedAt || previousEntry?.updatedAt || Date.now(),
-            hash,
-            deleted: false
-        };
-    }
-
-    const singletons = await collectSingletonEntries();
-    for (const item of singletons) {
-        if (item.data === null || item.data === undefined) continue;
-        const manifestKey = entryKey(item.type, item.id);
-        const hash = await computeSyncHash(item.data);
-        const previousEntry = previousEntries[manifestKey];
-        manifest.entries[entryKey(item.type, item.id)] = {
-            type: item.type,
-            id: item.id,
-            path: cloudPath(item.type, item.id),
-            updatedAt: previousEntry?.hash === hash && !previousEntry?.deleted
-                ? previousEntry.updatedAt
-                : Date.now(),
-            hash,
-            deleted: false
-        };
-    }
-
-    const deletions = await getSyncDeletedEntries();
-    for (const [key, deletion] of Object.entries(deletions)) {
-        manifest.entries[key] = {
-            type: deletion.type,
-            id: deletion.id,
-            path: cloudPath(deletion.type, deletion.id),
-            updatedAt: deletion.updatedAt,
-            hash: null,
-            deleted: true
-        };
-    }
-
-    manifest.lastSync = Date.now();
-    return manifest;
-}
-
-async function readCloudEntityByEntry(adapter, entry, key) {
-    let result = await adapter.download(entry.path);
-    if (!result) {
-        const altPath = entry.path.endsWith('.enc')
-            ? entry.path.replace('.enc', '.json')
-            : entry.path.replace('.json', '.enc');
-        result = await adapter.download(altPath);
-    }
-    if (!result) return null;
-    if (result.data.length > MAX_SYNC_PAYLOAD_BYTES) {
-        console.warn(`[sync] Skipping download ${entry.type} ${entry.id}: payload ${Math.round(result.data.length / 1024 / 1024)}MB exceeds limit`);
-        showToast(`Sync: ${entry.type} too large (${Math.round(result.data.length / 1024 / 1024)}MB), skipped`, 5000);
-        return null;
-    }
-    const parsed = JSON.parse(result.data);
-    return decryptEntity(parsed, key);
-}
-
-async function applyCloudEntry(adapter, entry, key) {
-    if (entry.deleted) {
-        if (entry.type === ENTITY_TYPES.CHARACTER) {
-            await db.delete('characters', entry.id);
-        } else if (entry.type === ENTITY_TYPES.PERSONA) {
-            await db.delete('personas', entry.id);
-        } else if (entry.type === ENTITY_TYPES.CHAT) {
-            await db.delete('keyvalue', `gz_chat_${entry.id}`);
-        }
-        await clearSyncDeletedEntry(entry.type, entry.id);
-        return null;
-    }
-
-    const entity = await readCloudEntityByEntry(adapter, entry, key);
-    if (entry.type === ENTITY_TYPES.CHARACTER) {
-        const existing = await db.get('characters', entry.id);
-        if (existing?.images) {
-            entity.images = existing.images;
-        } else if (!entity.images) {
-            entity.images = [];
-        }
-        await db.put('characters', entity);
-        await clearSyncDeletedEntry(entry.type, entry.id);
-    } else if (entry.type === ENTITY_TYPES.PERSONA) {
-        await db.put('personas', entity);
-        await clearSyncDeletedEntry(entry.type, entry.id);
-} else if (entry.type === ENTITY_TYPES.CHAT) {
-            await db.saveChat(entry.id, entity);
-            await clearSyncDeletedEntry(entry.type, entry.id);
-    } else if (entry.type === ENTITY_TYPES.LOREBOOKS) {
-        await db.set('gz_lorebooks', entity);
-    } else if (entry.type === ENTITY_TYPES.API_PRESETS) {
-        await db.set('gz_api_connection_presets', entity);
-    } else if (entry.type === ENTITY_TYPES.THEME_PRESETS) {
-        await db.set('gz_theme_presets', entity);
-    } else if (entry.type === ENTITY_TYPES.THEME_STATE) {
-        await db.set('gz_theme_active_preset', entity);
-    } else if (entry.type === ENTITY_TYPES.LOCAL_STORAGE) {
-        for (const [lsKey, lsVal] of Object.entries(entity || {})) {
-            if (lsKey === 'silly_cradle_presets') {
-                try {
-                    const cloudPresets = JSON.parse(lsVal);
-                    const localRaw = localStorage.getItem('silly_cradle_presets');
-                    const localPresets = localRaw ? JSON.parse(localRaw) : {};
-                    for (const [pId, cloudPreset] of Object.entries(cloudPresets)) {
-                        const localRegexes = localPresets[pId]?.regexes;
-                        const cloudRegexes = cloudPreset.regexes;
-                        if (localRegexes?.length) {
-                            if (!cloudRegexes?.length) {
-                                cloudPreset.regexes = localRegexes;
-                            } else {
-                                const cloudIds = new Set(cloudRegexes.map(r => r.id));
-                                const localOnly = localRegexes.filter(r => !cloudIds.has(r.id));
-                                if (localOnly.length) cloudPreset.regexes = [...cloudRegexes, ...localOnly];
-                            }
-                        }
-                    }
-                    localStorage.setItem(lsKey, JSON.stringify({ ...localPresets, ...cloudPresets }));
-                } catch (e) {
-                    localStorage.setItem(lsKey, lsVal);
-                }
-            } else if (lsKey === 'regex_scripts') {
-                try {
-                    const cloudScripts = JSON.parse(lsVal);
-                    const localRaw = localStorage.getItem('regex_scripts');
-                    const localScripts = localRaw ? JSON.parse(localRaw) : [];
-                    if (!Array.isArray(cloudScripts)) {
-                        localStorage.setItem(lsKey, lsVal);
-                    } else if (!localScripts.length) {
-                        localStorage.setItem(lsKey, lsVal);
-                    } else {
-                        const cloudIds = new Set(cloudScripts.map(r => r.id));
-                        const localOnly = localScripts.filter(r => !cloudIds.has(r.id));
-                        localStorage.setItem(lsKey, JSON.stringify([...cloudScripts, ...localOnly]));
-                    }
-                } catch (e) {
-                    localStorage.setItem(lsKey, lsVal);
-                }
-            } else {
-                localStorage.setItem(lsKey, lsVal);
-            }
-        }
-    }
-    return entity;
-}
-
-function getBreakdownBucket(type) {
-    if (type === ENTITY_TYPES.CHARACTER) return 'characters';
-    if (type === ENTITY_TYPES.PERSONA) return 'personas';
-    if (type === ENTITY_TYPES.CHAT) return 'chats';
-    if (type === ENTITY_TYPES.GALLERY) return 'gallery';
-    return 'settings';
-}
-
-function needsConflict(localEntry, cloudEntry) {
-    return !!localEntry && !localEntry.deleted && localEntry.updatedAt > cloudEntry.updatedAt;
-}
-
-async function getLocalConflictEntity(type, id) {
-    if (type === ENTITY_TYPES.CHARACTER) return getLocalCharacterWithImages(id);
-    if (type === ENTITY_TYPES.PERSONA) return getLocalPersona(id);
-    if (type === ENTITY_TYPES.CHAT) return getLocalChat(id);
-    return null;
-}
-
-function getConflictName(type, localEntity, cloudEntity, id) {
-    if (type === ENTITY_TYPES.CHARACTER || type === ENTITY_TYPES.PERSONA) {
-        return localEntity?.name || cloudEntity?.name || id;
-    }
-    if (type === ENTITY_TYPES.CHAT) {
-        return getChatName(localEntity, cloudEntity, id);
-    }
-    return id;
-}
-
-async function computeBinaryHash(arrayBuffer) {
-    const digest = await crypto.subtle.digest('SHA-256', arrayBuffer);
-    return Array.from(new Uint8Array(digest)).map(b => b.toString(16).padStart(2, '0')).join('');
-}
-
-async function applyGalleryEntriesForChar(adapter, entries, charId, localEntries, onConflict) {
-    const char = await getLocalCharacterWithImages(charId);
-    if (!char) return entries.length;
-    if (!Array.isArray(char.images)) char.images = [];
-
-    const localHashes = new Map();
-    for (const img of char.images) {
-        if (img?.src) {
-            try {
-                const h = await computeImageHash(img.src);
-                localHashes.set(h, img);
-            } catch {}
-        }
-    }
-
-    let applied = 0;
-    const galleryConflicts = [];
-    for (const cloudEntry of entries) {
-        const localEntry = localEntries?.[entryKey(ENTITY_TYPES.GALLERY, cloudEntry.id)];
-        if (localEntry && !localEntry.deleted && localEntry.updatedAt > cloudEntry.updatedAt) {
-            const conflict = {
-                type: ENTITY_TYPES.GALLERY,
-                id: cloudEntry.id,
-                name: `Gallery: ${cloudEntry.imgId}`,
-                charId,
-                imgId: cloudEntry.imgId,
-                local: { imgId: localEntry.imgId, hash: localEntry.hash, updatedAt: localEntry.updatedAt },
-                cloud: { imgId: cloudEntry.imgId, hash: cloudEntry.hash, updatedAt: cloudEntry.updatedAt },
-                cloudModified: cloudEntry.updatedAt
-            };
-            galleryConflicts.push(conflict);
-            if (onConflict) onConflict(conflict);
-            continue;
-        }
-
-        try {
-            if (cloudEntry.deleted) {
-                char.images = char.images.filter(im => im.id !== cloudEntry.imgId);
-                applied++;
-                continue;
-            }
-
-            const binary = await adapter.downloadBinary(cloudEntry.path);
-            if (!binary) continue;
-
-            const cloudHash = await computeBinaryHash(binary);
-            const existingByIdIdx = char.images.findIndex(im => im.id === cloudEntry.imgId);
-
-            if (existingByIdIdx >= 0) {
-                const blob = new Blob([binary]);
-                const dataUrl = await new Promise((resolve, reject) => {
-                    const reader = new FileReader();
-                    reader.onload = () => resolve(reader.result);
-                    reader.onerror = reject;
-                    reader.readAsDataURL(blob);
-                });
-                const old = char.images[existingByIdIdx];
-                char.images[existingByIdIdx] = { id: cloudEntry.imgId, src: dataUrl, name: old?.name || '', thumbnail: null };
-                applied++;
-            } else {
-                const existingByHash = localHashes.get(cloudHash);
-                if (existingByHash) {
-                    const existingIdx = char.images.findIndex(im => im.id === existingByHash.id);
-                    const blob = new Blob([binary]);
-                    const dataUrl = await new Promise((resolve, reject) => {
-                        const reader = new FileReader();
-                        reader.onload = () => resolve(reader.result);
-                        reader.onerror = reject;
-                        reader.readAsDataURL(blob);
-                    });
-                    if (existingIdx >= 0) {
-                        char.images[existingIdx] = { id: cloudEntry.imgId, src: dataUrl, name: char.images[existingIdx]?.name || '', thumbnail: null };
-                    } else {
-                        char.images.push({ id: cloudEntry.imgId, src: dataUrl, name: '', thumbnail: null });
-                    }
-                    localHashes.delete(cloudHash);
-                    localHashes.set(cloudHash, { id: cloudEntry.imgId });
-                    applied++;
-                } else {
-                    const blob = new Blob([binary]);
-                    const dataUrl = await new Promise((resolve, reject) => {
-                        const reader = new FileReader();
-                        reader.onload = () => resolve(reader.result);
-                        reader.onerror = reject;
-                        reader.readAsDataURL(blob);
-                    });
-                    char.images.push({ id: cloudEntry.imgId, src: dataUrl, name: '', thumbnail: null });
-                    localHashes.set(cloudHash, { id: cloudEntry.imgId });
-                    applied++;
-                }
-            }
-        } catch (e) {
-            console.warn(`[sync] Gallery entry ${cloudEntry.imgId} for char ${charId} failed:`, e);
-        }
-    }
-
-    await db.put('characters', char);
-    return { applied, conflicts: galleryConflicts };
-}
-
-async function removeGalleryEntryFromChar(charId, imgId) {
-    const char = await getLocalCharacterWithImages(charId);
-    if (char?.images) {
-        char.images = char.images.filter(im => im.id !== imgId);
-        await db.put('characters', char);
-    }
-}
-
-async function deleteCloudFileIfExists(adapter, entry) {
-    try {
-        await adapter.deleteFile(entry.path);
-    } catch {
-        // Ignore missing payload deletes; manifest remains source of truth.
-    }
 }
 
 async function runParallel(tasks, concurrency = 5, delayMs = 200) {
@@ -666,12 +122,14 @@ async function runParallel(tasks, concurrency = 5, delayMs = 200) {
     return results;
 }
 
+const _deps = { computeSyncHash, cloudPath, galleryCloudPath, computeImageHash, guessImageExt, entryKey, ENTITY_TYPES, getLocalCharacterWithImages, getLocalPersona, getLocalChat, listAllFiles };
+
 async function pushManifestV2(adapter, key, onProgress) {
     if (adapter.ensureFolder) {
         await adapter.ensureFolder(CLOUD_BASE);
     }
-    const cloudManifest = await readCloudManifestV2(adapter);
-    const localManifest = await buildLocalManifestV2();
+    const cloudManifest = await readCloudManifestV2(adapter, _deps);
+    const localManifest = await buildLocalManifestV2(_deps);
     const cloudEntries = cloudManifest?.entries || {};
     const breakdown = { characters: 0, personas: 0, chats: 0, gallery: 0, settings: 0 };
     let pushed = 0;
@@ -760,7 +218,6 @@ async function pushManifestV2(adapter, key, onProgress) {
                     const serialized = JSON.stringify(encrypted);
                     if (serialized.length > MAX_SYNC_PAYLOAD_BYTES) {
                         console.warn(`[sync] Skipping ${localEntry.type} ${localEntry.id}: payload ${Math.round(serialized.length / 1024 / 1024)}MB exceeds ${Math.round(MAX_SYNC_PAYLOAD_BYTES / 1024 / 1024)}MB limit`);
-                        showToast(`${localEntry.type} too large to sync (${Math.round(serialized.length / 1024 / 1024)}MB), skipped`, 5000);
                         skipped++;
                         if (onProgress) onProgress(phase, i + 1, allEntries.length);
                         return;
@@ -788,7 +245,7 @@ async function pushManifestV2(adapter, key, onProgress) {
 async function pullManifestV2(adapter, key, onProgress, onConflict) {
     let cloudManifest;
     try {
-        cloudManifest = await readCloudManifestV2(adapter);
+        cloudManifest = await readCloudManifestV2(adapter, _deps);
     } catch (e) {
         throw new Error(`Cannot access cloud data: ${e.message}`);
     }
@@ -796,7 +253,7 @@ async function pullManifestV2(adapter, key, onProgress, onConflict) {
         throw new Error(`Cloud manifest not found. No sync data was found in the cloud folder. If this is a new device, try pushing your local data first.`);
     }
 
-    const localManifest = await buildLocalManifestV2();
+    const localManifest = await buildLocalManifestV2(_deps);
     const cloudEntries = cloudManifest.entries || {};
     const localEntries = localManifest.entries || {};
     const allKeys = new Set([...Object.keys(cloudEntries), ...Object.keys(localEntries)]);
@@ -832,7 +289,7 @@ async function pullManifestV2(adapter, key, onProgress, onConflict) {
     }
 
     for (const [charId, entries] of galleryEntriesByChar) {
-        const result = await applyGalleryEntriesForChar(adapter, entries, charId, localEntries, onConflict);
+        const result = await applyGalleryEntriesForChar(adapter, entries, charId, localEntries, onConflict, { entryKey, ENTITY_TYPES });
         pulled += result.applied;
         breakdown.gallery += result.applied;
         if (result.conflicts.length > 0) {
@@ -846,7 +303,7 @@ async function pullManifestV2(adapter, key, onProgress, onConflict) {
     const ngTasks = nonGalleryTasks.map(({ cloudEntry, localEntry, phase, index }) => {
         return async () => {
             if (needsConflict(localEntry, cloudEntry)) {
-                const localEntity = await getLocalConflictEntity(cloudEntry.type, cloudEntry.id);
+                const localEntity = await getLocalConflictEntity(cloudEntry.type, cloudEntry.id, _deps);
                 let cloudEntity = null;
                 if (!cloudEntry.deleted) {
                     try {
@@ -907,18 +364,6 @@ export async function pullEntities(adapter, key, onProgress, onConflict) {
     return pullManifestV2(adapter, key, onProgress, onConflict);
 }
 
-async function readManifest(adapter) {
-    try {
-        const result = await adapter.download(cloudPath(ENTITY_TYPES.MANIFEST));
-        if (result) {
-            return JSON.parse(result.data);
-        }
-        return null;
-    } catch {
-        return null;
-    }
-}
-
 async function listAllFiles(adapter) {
     const files = [];
     try {
@@ -955,36 +400,6 @@ async function listAllFiles(adapter) {
         console.warn('[syncEngine] listAllFiles error:', e);
     }
     return files;
-}
-
-async function getLocalCharacter(id) {
-    const char = await db.get('characters', id);
-    if (!char) return null;
-    const { images: _images, ...rest } = char;
-    return rest;
-}
-
-async function getLocalCharacterWithImages(id) {
-    return (await db.get('characters', id)) || null;
-}
-
-async function getLocalPersona(id) {
-    return (await db.get('personas', id)) || null;
-}
-
-async function getLocalChat(charId) {
-    return db.getChat(charId);
-}
-
-function getChatName(localChat, cloudChat, charId) {
-    const msgs = cloudChat?.messages || localChat?.messages || [];
-    if (msgs.length > 0) {
-        const first = msgs[0];
-        const text = first.mes || first.content || '';
-        const preview = text.substring(0, 40).replace(/\n/g, ' ');
-        return preview || charId;
-    }
-    return charId;
 }
 
 const WIPE_POLL_INTERVAL = 2000;
@@ -1028,35 +443,5 @@ export async function wipeCloudData(adapter, onProgress) {
 }
 
 export async function resolveConflict(conflict, choice) {
-    if (conflict.type === ENTITY_TYPES.GALLERY) {
-        if (choice === 'local') return null;
-        const char = await getLocalCharacterWithImages(conflict.charId);
-        if (!char) return null;
-        return char;
-    }
-    const entity = choice === 'cloud' ? conflict.cloud : conflict.local;
-    if (choice === 'cloud' && !entity) {
-        if (conflict.type === ENTITY_TYPES.CHARACTER) {
-            await db.delete('characters', conflict.id);
-        } else if (conflict.type === ENTITY_TYPES.PERSONA) {
-            await db.delete('personas', conflict.id);
-        } else if (conflict.type === ENTITY_TYPES.CHAT) {
-            await db.delete('keyvalue', `gz_chat_${conflict.id}`);
-        }
-        return null;
-    }
-    if (conflict.type === ENTITY_TYPES.CHARACTER) {
-        const existing = await db.get('characters', conflict.id);
-        if (existing?.images && !entity.images) {
-            entity.images = existing.images;
-        }
-        await db.put('characters', entity);
-    } else if (conflict.type === ENTITY_TYPES.PERSONA) {
-        await db.put('personas', entity);
-    } else if (conflict.type === ENTITY_TYPES.CHAT) {
-        await db.saveChat(conflict.id, entity);
-    }
-    return entity;
+    return _resolveConflict(conflict, choice, { ENTITY_TYPES, getLocalCharacterWithImages });
 }
-
-export { ENTITY_TYPES, CLOUD_BASE, cloudPath, getDeviceId };
